backend/requirements.txt
Django==4.0.3
djangorestframework==3.13.1
mysqlclient==2.1.0
django-mysql==4.5.0